--- packages/components/src/components/ui/atoms/Combobox/Combobox.stories.tsx
@@ -1,17 +1,0 @@
-import { Meta, StoryFn } from '@storybook/react'
-
-import Example from './Combobox'
-
-export default {
-  title: 'Atoms/Comobobox',
-  component: Example,
-  argTypes: {},
-} as Meta<typeof Example>
-
-const Template: StoryFn<{
-  subText?: string
-}> = () => {
-  return <Example />
-}
-
-export const Showcase = Template.bind({})
